+++ deltafi-ui/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Integrated default external links into core menus (#269)
 - Added support for displaying Indexed Metadata on DeltaFile Viewer page (#252)
 - Added New Dashboard Panels (#265)
+- Added Import and Rollback capabilities to System Snapshots page (#259)
 
 ### Fixed